refactor(transfer): split receiver into submodules
